+++ src/helpers/apiHelpers.js
@@ -307,119 +307,6 @@ export async function downloadManifestData(owner, repo, retries=5) {
   }
 }
 
-/**
- * reads files names from folder and remove system files
- * @param {string} path
- * @return {*[]}
- */
-export const cleanReaddirSync = (path) => {
-  let cleanDirectories = [];
-
-  if (fs.existsSync(path)) {
-    cleanDirectories = fs.readdirSync(path)
-      .filter((file) => file !== '.DS_Store');
-  } else {
-    console.warn(`no such file or directory, ${path}`);
-  }
-
-  return cleanDirectories;
-};
-
-/**
- * takes path that has version and owner such as `~/resources/bible/v1.1_unfoldingWord` and extracts the version and owner.
- * @param {string} versionPath
- * @return {{owner: string, version: string}}
- */
-export function getVersionAndOwnerFromPath(versionPath) {
-  if (versionPath) {
-    const versionAndOwner = path.base(versionPath, true);
-    return splitVersionAndOwner(versionAndOwner);
-  }
-  return {};
-}
-
-/**
- * takes string that has version and owner such as `v1.1_unfoldingWord` and splits into version and owner.
- * @param {string} versionAndOwner
- * @return {{owner: string, version: string}}
- */
-export function splitVersionAndOwner(versionAndOwner) {
-  let version = versionAndOwner;
-  let owner = '';
-  const pos = versionAndOwner.indexOf(OWNER_SEPARATOR);
-  if (pos >= 0) {
-    owner = decodeURIComponent(versionAndOwner.substr(pos + 1));
-    version = versionAndOwner.substr(0, pos);
-  }
-  return {version, owner};
-}
-
-/**
- * Returns the versioned folder within the directory with the highest value.
- * e.g. `v10` is greater than `v9`
- * @param {string} dir - the directory to read
- * @return {string} the full path to the latest version directory.
- */
-export function getLatestVersionsAndOwners(dir) {
-  const versionAndOwners = listVersions(dir);
-  const orgs = {};
-
-  for (const versionAndOwner of versionAndOwners) {
-    let {owner} = splitVersionAndOwner(versionAndOwner);
-    if (!orgs[owner]) {
-      orgs[owner] = [];
-    }
-    orgs[owner].push(versionAndOwner);
-  }
-
-  const orgsKeys = Object.keys(orgs);
-  for (const org of orgsKeys) {
-    const versions = orgs[org];
-    const latest = path.join(dir, versions[0]);
-    orgs[org] = latest;
-  }
-
-  if (orgsKeys.length > 0) {
-    return orgs;
-  } else {
-    return null;
-  }
-}
-
-/**
- * Returns an array of paths found in the directory filtered and sorted by version
- * @param {string} dir
- * @return {string[]}
- */
-export function listVersions(dir) {
-  if (fs.pathExistsSync(dir)) {
-    const versionedDirs = fs.readdirSync(dir).filter((file) => fs.lstatSync(path.join(dir, file)).isDirectory() &&
-      file.match(/^v\d/i));
-    return versionedDirs.sort((a, b) =>
-      -compareVersions(a, b), // do inverted sort
-    );
-  }
-  return [];
-}
-
-/**
- * compares version numbers, if a > b returns 1; if a < b return -1; else are equal and return 0
- * @param {String} a
- * @param {String} b
- * @return {number}
- */
-export function compareVersions(a, b) {
-  const cleanA = semver.coerce(a);
-  const cleanB = semver.coerce(b);
-
-  if (semver.gt(cleanA, cleanB)) {
-    return 1;
-  } else if (semver.lt(cleanA, cleanB)) {
-    return -1;
-  } else {
-    return 0;
-  }
-}
 
 /**
  * get local resources
